Fix SwiftLint file_length and type_body_length violations in ModelSearch files
