+++ iosApp/iosApp/Features/Search/ModelSearchViewModel.swift
@@ -254,78 +254,6 @@ final class ModelSearchViewModel: ObservableObject {
         }
     }
 
-    func toggleFavorite(_ model: Model) {
-        Task {
-            try? await toggleFavoriteUseCase.invoke(model: model)
-        }
-    }
-    func observeFavorites() async {
-        for await list in observeFavoritesUseCase.invoke() {
-            let summaries = list.compactMap { $0 as? FavoriteModelSummary }
-            favoriteIds = Set(summaries.map { $0.id })
-        }
-    }
-
-    func observeSavedFilters() async {
-        for await list in observeSavedSearchFiltersUseCase.invoke() {
-            savedFilters = list.compactMap { $0 as? SavedSearchFilter }
-        }
-    }
-
-    func saveCurrentFilter(name: String) {
-        let filter = SavedSearchFilter(
-            id: 0,
-            name: name,
-            query: query,
-            selectedType: selectedType,
-            selectedSort: selectedSort,
-            selectedPeriod: selectedPeriod,
-            selectedBaseModels: selectedBaseModels,
-            nsfwFilterLevel: nsfwFilterLevel,
-            isFreshFindEnabled: isFreshFindEnabled,
-            excludedTags: excludedTags,
-            includedTags: includedTags,
-            savedAt: 0
-        )
-        Task { try? await saveSearchFilterUseCase.invoke(name: name, filter: filter) }
-    }
-
-    func applyFilter(_ filter: SavedSearchFilter) {
-        loadTask?.cancel()
-        query = filter.query
-        selectedType = filter.selectedType
-        selectedSort = filter.selectedSort
-        selectedPeriod = filter.selectedPeriod
-        selectedBaseModels = Set(filter.selectedBaseModels.compactMap { $0 as? BaseModel })
-        nsfwFilterLevel = filter.nsfwFilterLevel
-        isFreshFindEnabled = filter.isFreshFindEnabled
-        includedTags = filter.includedTags.compactMap { $0 as? String }
-        excludedTags = filter.excludedTags.compactMap { $0 as? String }
-        reloadModels()
-    }
-
-    func deleteSavedFilter(id: Int64) {
-        Task { try? await deleteSavedSearchFilterUseCase.invoke(id: id) }
-    }
-    func hideModel(_ modelId: Int64, name: String) {
-        Task {
-            try? await hideModelUseCase.invoke(modelId: modelId, modelName: name)
-            hiddenModelIds = try await getHiddenModelIdsUseCase.invoke()
-            models.removeAll { $0.id == modelId }
-        }
-    }
-    func isModelOwned(_ model: Model) -> Bool {
-        guard !ownedHashes.isEmpty else { return false }
-        return model.modelVersions.contains { version in
-            version.files.contains { file in
-                if let sha256 = file.hashes["SHA256"] as? String {
-                    return ownedHashes.contains(sha256.lowercased())
-                }
-                return false
-            }
-        }
-    }
-
     private func loadExcludedTags() {
         Task {
             excludedTags = (try? await getExcludedTagsUseCase.invoke()) ?? []
@@ -498,52 +426,65 @@ final class ModelSearchViewModel: ObservableObject {
     }
 }
 
-// MARK: - NSFW Image Filtering
+// MARK: - Favorites, Hidden Models & Saved Filters
 
-extension ModelImage {
-    func isAllowedByFilter(_ level: NsfwFilterLevel) -> Bool {
-        switch level {
-        case .off: return nsfwLevel == .none
-        case .soft: return nsfwLevel == .none || nsfwLevel == .soft
-        default: return true
+extension ModelSearchViewModel {
+    func toggleFavorite(_ model: Model) {
+        Task { try? await toggleFavoriteUseCase.invoke(model: model) }
+    }
+    func observeFavorites() async {
+        for await list in observeFavoritesUseCase.invoke() {
+            let summaries = list.compactMap { $0 as? FavoriteModelSummary }
+            favoriteIds = Set(summaries.map { $0.id })
         }
     }
-}
-
-extension Array where Element == Model {
-    func filterNsfwImages(_ level: NsfwFilterLevel) -> [Model] {
-        if level == .all { return self }
-        return compactMap { model in
-            let filteredVersions = model.modelVersions.map { version in
-                let safeImages = version.images.filter { $0.isAllowedByFilter(level) }
-                return ModelVersion(
-                    id: version.id,
-                    modelId: version.modelId,
-                    name: version.name,
-                    description: version.description_,
-                    createdAt: version.createdAt,
-                    baseModel: version.baseModel,
-                    trainedWords: version.trainedWords,
-                    downloadUrl: version.downloadUrl,
-                    files: version.files,
-                    images: safeImages,
-                    stats: version.stats
-                )
-            }
-            let hasAnyImages = filteredVersions.contains { !$0.images.isEmpty }
-            guard hasAnyImages else { return nil }
-            return Model(
-                id: model.id,
-                name: model.name,
-                description: model.description_,
-                type: model.type,
-                nsfw: model.nsfw,
-                tags: model.tags,
-                mode: model.mode,
-                creator: model.creator,
-                stats: model.stats,
-                modelVersions: filteredVersions
-            )
+    func hideModel(_ modelId: Int64, name: String) {
+        Task {
+            try? await hideModelUseCase.invoke(modelId: modelId, modelName: name)
+            hiddenModelIds = try await getHiddenModelIdsUseCase.invoke()
+            models.removeAll { $0.id == modelId }
         }
+    }
+    func isModelOwned(_ model: Model) -> Bool {
+        guard !ownedHashes.isEmpty else { return false }
+        return model.modelVersions.contains { version in
+            version.files.contains { file in
+                if let sha256 = file.hashes["SHA256"] as? String {
+                    return ownedHashes.contains(sha256.lowercased())
+                }
+                return false
+            }
+        }
+    }
+    func observeSavedFilters() async {
+        for await list in observeSavedSearchFiltersUseCase.invoke() {
+            savedFilters = list.compactMap { $0 as? SavedSearchFilter }
+        }
+    }
+    func saveCurrentFilter(name: String) {
+        let filter = SavedSearchFilter(
+            id: 0, name: name, query: query, selectedType: selectedType,
+            selectedSort: selectedSort, selectedPeriod: selectedPeriod,
+            selectedBaseModels: selectedBaseModels, nsfwFilterLevel: nsfwFilterLevel,
+            isFreshFindEnabled: isFreshFindEnabled, excludedTags: excludedTags,
+            includedTags: includedTags, savedAt: 0
+        )
+        Task { try? await saveSearchFilterUseCase.invoke(name: name, filter: filter) }
+    }
+    func applyFilter(_ filter: SavedSearchFilter) {
+        loadTask?.cancel()
+        query = filter.query
+        selectedType = filter.selectedType
+        selectedSort = filter.selectedSort
+        selectedPeriod = filter.selectedPeriod
+        selectedBaseModels = Set(filter.selectedBaseModels.compactMap { $0 as? BaseModel })
+        nsfwFilterLevel = filter.nsfwFilterLevel
+        isFreshFindEnabled = filter.isFreshFindEnabled
+        includedTags = filter.includedTags.compactMap { $0 as? String }
+        excludedTags = filter.excludedTags.compactMap { $0 as? String }
+        reloadModels()
+    }
+    func deleteSavedFilter(id: Int64) {
+        Task { try? await deleteSavedSearchFilterUseCase.invoke(id: id) }
     }
 }
